requirements.txt
Pillow==6.2.1
selenium==3.141.0
urllib3==1.25.6
